Implement skeleton for services and repositories

+++ internal/domain/review.go
@@ -5,10 +5,10 @@ import (
 )
 
 type Review struct {
-	ID          int       `json:"id"`
-	GameID      int       `json:"gameID"`
-	UserID      int       `json:"userID"`
-	Rating      int       `json:"rating"`
+	ID          int64       `json:"id"`
+	GameID      int64       `json:"gameID"`
+	UserID      int64       `json:"userID"`
+	Rating      int8       `json:"rating"`
 	ReviewText  string    `json:"reviewText"`
 	DateCreated time.Time `json:"dateCreated"`
 }
